FreeEmsPlugin: Better handling of disconnect detection and reconnect syncing
EMSTune will now pop up a display when it detects that it has connected to
a device with differing non-readonly memory locations. It will also allow the
user to either send local changes to the ECU,or accept ECU defaults as the
current memory situation.

The plugin will also notify the user when it has detected that the ECU has
gone silent for more than 1500 milliseconds. This will be user configurable
in the future once plugin-specific settings are supported.

Added functionality to actually check if incoming data is
different from what is stored in application memory and
sendinga notification to the the user. This is a temp commit
to trigger a test build and will be replaced with an actual
commit containing a proper diff window of some sort of the data

iTEmp